refactor: extract shared users logic into monorepo library
