+++ libs/users/src/constants/collections.constant.ts
@@ -1,0 +1,3 @@
+export const USERS_MODULE_COLLECTIONS = {
+    User: 'User',
+};
