--- src/app/users/constants/collections.constant.ts
@@ -1,3 +1,0 @@
-export const COLLECTIONS = {
-    User: 'User',
-};
